% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASiCS_ParallelMCMC.R
\name{BASiCS_ParallelMCMC}
\alias{BASiCS_ParallelMCMC}
\title{Parallelized BASiCS_MCMC}
\usage{
BASiCS_ParallelMCMC(
  CountsList,
  SpikeInfo,
  N = 200000L,
  Burn = 80000L,
  Thin = 40L,
  Regression = TRUE,
  MinCount = 10L,
  PropMinCount = 0.85,
  NCores = 2L,
  Name = paste0(sample(c(0:9, letters), 10), collapse = ""),
  Verbose = TRUE
)
}
\arguments{
\item{CountsList}{A list of unfiltered count matrices with genes as rows and samples as columns.}

\item{SpikeInfo}{\code{data.frame} whose first and second columns contain
the gene names assigned to the spike-in genes (they must match the ones in
\code{rownames(Counts)}) and the associated input number of molecules,
respectively. If \code{SpikeInfo = NULL}, only the horizontal integration
implementation (no spikes) can be run. Default value: \code{SpikeInfo = NULL}.}

\item{N}{Total number of iterations for the MCMC sampler.
Use \code{N>=max(4,Thin)}, \code{N} being a multiple of \code{Thin}.}

\item{Burn}{Burn-in period for the MCMC sampler. Use \code{Burn>=1},
\code{Burn<N}, \code{Burn} being a multiple of \code{Thin}.}

\item{Thin}{Thining period for the MCMC sampler. Use \code{Thin>=2}.}

\item{Regression}{If \code{Regression = TRUE}, BASiCS exploits a joint prior
formulation for mean and over-dispersion parameters to estimate a measure of
residual over-dispersion is not confounded by mean expression. Recommended
setting is \code{Regression = TRUE}.}

\item{MinCount}{integer; Minimum number of counts for a gene.}

\item{Name}{character; Chain identification name.}

\item{PropMinCount;}{numeric; Proportion of samples with minimum count.}

\item{NCores;}{integer; Number of cores.}

\item{Verbose;}{bool; Print progress in \code{./logs} directory?}
}
\value{
A list of \link[BASiCS]{BASiCS_Chain} object.
}
\description{
Parallelize BASiCS normalization implemented in \link[BASiCS]{BASiCS_MCMC}.
}
\details{
Assumes that there are spike-ins and that these are named "ERCC".
Filter rule: a gene must be counted at least \code{MinCount} in \code{PropMinCount} percent of the samples.
For biological transcripts, this needs to be true in at least one group. For spike-ins in all the groups.
Uses \link[parallel]{mclapply} to parallelize.
}
\seealso{
\link[BASiCS]{BASiCS_Chain}

\link[BASiCS]{BASiCS_MCMC}
}
