% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castToMatrix.R
\name{castToMatrix}
\alias{castToMatrix}
\title{Cast a data.table to a matrix}
\usage{
castToMatrix(data, formula, fun.aggregate, value.var, ...)
}
\arguments{
\item{data}{ A \code{data.table}.}

\item{formula}{A formula of the form LHS ~ RHS to cast, see Details.}

\item{fun.aggregate}{Should the data be aggregated before casting? If the formula doesn't identify a single observation for each cell, then aggregation defaults to \code{length} with a message.

  To use multiple aggregation functions, pass a \code{list}; see Examples. }

\item{value.var}{Name of the column whose values will be filled to cast. Function \code{guess()} tries to, well, guess this column automatically, if none is provided.

  Cast multiple \code{value.var} columns simultaneously by passing their names as a \code{character} vector. See Examples. }

\item{...}{Any other arguments that may be passed to the aggregating function.}
}
\value{
A matrix with columns and rows corresponding to the formula.
}
\description{
A wrap-around the \code{dcast} function that returns a \code{matrix} instead of a \code{data.table}
}
\seealso{
\link[data.table]{dcast}
}
