% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotHeatmap.R
\name{clusteredCorrplot}
\alias{clusteredCorrplot}
\title{Correlation plot with dendograms}
\usage{
clusteredCorrplot(x, hc, k = NULL, dendogram_colors = NULL)
}
\arguments{
\item{x}{data matrix}

\item{hc}{hierarchical clustering (output of \link{hclust})}

\item{k}{number of clusters}

\item{dendogram_colors}{cluster colors in dendogram (length should be k)}
}
\description{
Plots a correlation plot with dendograms using ggplot2
}
\examples{
hc <- hclust(as.dist(1-cor(trees)))
clusteredCorrplot(trees, hc, 2, c("red", "blue"))

}
