% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGeneNames.R
\name{convertGene}
\alias{convertGene}
\title{Convert gene names}
\usage{
convertGene(keys, column = "SYMBOL", keytype = "WORMBASE")
}
\arguments{
\item{keys}{the keys to select records for from the database.  All possible 
    keys are returned by using the \code{keys} method.}

\item{column}{the column to search on (for \code{mapIds}).  Different from 
    \code{columns} in that it can only have a single element for the value}

\item{keytype}{the keytype that matches the keys used.  For the
    \code{select} methods, this is used to indicate the kind of ID being used
    with the keys argument. For the \code{keys} method this is used to
    indicate which kind of keys are desired from \code{keys}
  }
}
\value{
character; Converted gene identification to the type specified by \code{column}.
}
\description{
Convert wormbase gene identification to official gene symbol. Uses \code{org.Ce.eg.db} for mapping.
}
\seealso{
\link[AnnotationDbi]{mapIds}
}
