% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctedCountsSVA.R
\name{correctedCountsSVA}
\alias{correctedCountsSVA}
\title{Matrix of batch corrected expression values}
\usage{
correctedCountsSVA(y, mod, svaobj)
}
\arguments{
\item{y}{matrix; Counts}

\item{mod}{The model matrix being used to fit the data}

\item{svaseq}{\link[sva]{svaseq} object}
}
\value{
A matrix of batch corrected counts.
}
\description{
Computes batch corrected counts using SVA.
}
\seealso{
\link[sva]{svaseq}
}
