% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutree2.R
\name{cutree2}
\alias{cutree2}
\title{Cut a Tree into Groups of Data}
\usage{
cutree2(tree, k = NULL, h = NULL)
}
\arguments{
\item{tree}{a tree as produced by \code{\link[stats]{hclust}}. \code{cutree()}
   only expects a list with components \code{merge}, \code{height}, and
   \code{labels}, of appropriate content each.}

\item{k}{an integer scalar or vector with the desired number of groups}

\item{h}{numeric scalar or vector with heights where the tree should
   be cut.}
}
\description{
Does the same as \link[stats]{cutree}, but the numbering
matches the order of clusters when plotting the tree
produced by \link[stats]{hclust}
}
