% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupSparseEigen.R
\name{groupSparseEigen}
\alias{groupSparseEigen}
\title{Group-wise sparse spectral decomposition of a matrix}
\usage{
groupSparseEigen(
  x,
  groups,
  y = NULL,
  lambda = NULL,
  kernel = "linear",
  bandwidth = NULL,
  trace = FALSE
)
}
\arguments{
\item{x}{numeric matrix; data with variables as columns and samples as rows.}

\item{groups}{boolean matrix; affiliation matrix. Does the variable in row i pertain to the group in column j?}

\item{y}{numeric matrix; confounder matrix with rows as samples and columns as confounding factors.}

\item{lambda}{the tuning parameter, non-negative.}

\item{kernel}{the kernel to use: "linear", "gaussian".}

\item{bandwidth}{bandwidth h for Gaussian kernel. Optional.}

\item{trace}{logical; should progress be printed?}
}
\value{
A list containing the sparse eigenvector and eigenvalues.
}
\description{
Computes group-wise sparse eigenvalues and eigenvectors of a matrix.
}
\details{
For each group defined in the \code{groups} parameter,
variables not pertaining to the group see their covariance set to zero.
Eigendecomposition is then performed on that matrix.
}
\seealso{
\link[base]{eigen}
}
