% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kFoldRsq.R
\name{kFoldRsq}
\alias{kFoldRsq}
\title{Cross validated R2 from linear models}
\usage{
kFoldRsq(lmfit, ngroup = 10)
}
\arguments{
\item{lmfit}{object of class \code{lm}}

\item{ngroup}{optional argument specifying the number of  groups formed .
       Default is \code{ngroup}=sample size, corresponding to leave-one out
       cross-validation.}
}
\value{
Raw and cross-validated R2 and adjusted R2
}
\description{
Estimates out of sample R2 and adjusted R2 using using cross-validation
}
