% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\alias{overlapNormal}
\alias{overlapLogNormal}
\title{Overlapping estimation}
\usage{
overlap(x, y, nbins = 1024, ...)

overlapNormal(x, mean, sd, nbins = 1024, ...)

overlapLogNormal(x, meanlog, sdlog, nbins = 1024, ...)
}
\arguments{
\item{x}{numerical vector}

\item{y}{numerical vector}

\item{nbins}{number of equally spaced points at which the overlapping density is evaluated see density for details}

\item{...}{optional arguments to be passed to function density}

\item{mean}{mean of normal distribution}

\item{meanlog}{mean of log-normal distribution in log scale}

\item{sdlog}{standard deviation of log-normal distribution in log scale}

\item{sigma}{standard deviation of normal distribution}
}
\value{
Estimates of overlapped areas relative to each pair of distributions
}
\description{
Estimates overlapping area of two distributions.
}
\section{Functions}{
\itemize{
\item \code{overlap}: 

\item \code{overlapNormal}: 

\item \code{overlapLogNormal}: 
}}

\seealso{
\link[overlapping]{overlap}
}
