% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentAnalysis.R
\name{plotGSEA}
\alias{plotGSEA}
\title{Plot GSEA results}
\usage{
plotGSEA(tables, showCategory = 50L)
}
\arguments{
\item{tables}{List of \code{enrichResult} produced by \link{runGSEA}.}
}
\value{
An object of class ggarrange, which is a ggplot or a list of ggplot.
}
\description{
Plot results produced by \link{runGSEA}.
}
