% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverselog_trans.R
\name{reverselog_trans}
\alias{reverselog_trans}
\title{Inverse log scale}
\usage{
reverselog_trans(base = 10)
}
\arguments{
\item{base}{numeric; logarithm base.}
}
\description{
Transform data in the inverse log scale for ggplot2
}
\examples{
\donttest{
ggplot(trees, aes(Girth, Height)) +
  geom_point() +
  scale_y_continuous(trans=reverselog_trans())
  }

}
\seealso{
\link[ggplot2]{scale_continuous}
}
