% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpca.R
\name{rpca}
\alias{rpca}
\title{RPCA}
\usage{
rpca(
  x,
  k,
  scale = FALSE,
  sign_correction = TRUE,
  rotation = c("none", "varimax", "oblimin"),
  normalize = TRUE,
  maxit = 1000,
  eps = 1e-05,
  return_data = FALSE
)
}
\arguments{
\item{x}{numeric matrix; data with variables as columns and samples as rows.}

\item{k}{integer; number of principal components.}

\item{scale}{logical; should the data be scaled?}

\item{sign_correction}{logical; should \link{signCorrectionPCA} be called?}

\item{normalize}{parameter passed to optimization routine (GPForth or GPFoblq).}

\item{maxit}{parameter passed to optimization routine (GPForth or GPFoblq).}

\item{eps}{parameter passed to optimization routine (GPForth or GPFoblq).}

\item{return_data}{logical; should data be returned?}
}
\value{
A list with scores and loadings.
}
\description{
Rotated principal component analysis
}
