% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentAnalysis.R
\name{runEnrichment}
\alias{runEnrichment}
\title{Run enrichment analysis of a gene set for C. elegans}
\usage{
runEnrichment(
  gene,
  universe = NULL,
  keytype = "WORMBASE",
  pvalueCutoff = 1,
  use_internal_data = TRUE
)
}
\arguments{
\item{gene}{a vector of entrez gene id.}

\item{universe}{background genes. If missing, the all genes listed in the database (eg TERM2GENE table) will be used as background.}

\item{keytype}{the keytype that matches the keys used.  For the
    \code{select} methods, this is used to indicate the kind of ID being used
    with the keys argument. For the \code{keys} method this is used to
    indicate which kind of keys are desired from \code{keys}
  }

\item{pvalueCutoff}{pvalue cutoff on enrichment tests to report}

\item{use_internal_data}{logical, use KEGG.db or latest online KEGG data}
}
\value{
A list of \code{enrichResult} for KEGG and GO annotations.
}
\description{
Wrapper function around \link[clusterProfiler]{enrichKEGG} and \link[clusterProfiler]{enrichGO}.
}
