% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signCorrectionPCA.R
\name{signCorrectionPCA}
\alias{signCorrectionPCA}
\title{Sign correction for principal component analysis}
\usage{
signCorrectionPCA(data, loadings, scores)
}
\arguments{
\item{data}{matrix; The data matrix with samples in the rows and variables in the columns.}

\item{loadings}{matrix; The loading matrix with variables in the rows and components in the columns.}

\item{scores}{matrix; The score matrix with samples in the rows and components in the columns.}
}
\value{
A list containing the sign corrected loadings and scores.
}
\description{
Corrects the sign ambiguity that exists in PCA by correlating scores with the data.
}
\details{
The sign of the jth component is defined as:
\code{sum(cor(scores[, j], data) * loadings[, j]**2))}
}
\seealso{
\link[stats]{prcomp}
}
