% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_to_tpm.R
\name{sleuth_to_tpm}
\alias{sleuth_to_tpm}
\title{Convert a sleuth object to TPM}
\usage{
sleuth_to_tpm(obj, log = TRUE, pc = 0.5)
}
\arguments{
\item{obj}{\code{sleuth} object.}

\item{log}{bool; should natural log be returned?}

\item{pc}{numeric; pseudo-counts}
}
\value{
a matrix which contains a matrix of target_ids and transcript (or gene) expression in TPMs.
}
\description{
Convert a sleuth object to a TPM matrix with the condition names.
}
