% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpca.R
\name{gpca}
\alias{gpca}
\title{GPCA}
\usage{
gpca(
  x,
  groups,
  scale = FALSE,
  sign_correction = TRUE,
  return_data = FALSE,
  trace = FALSE
)
}
\arguments{
\item{x}{numeric matrix; data with variables as columns and samples as rows.}

\item{groups}{boolean matrix; affiliation matrix. Does the variable in row i pertain to the group in column j?}

\item{scale}{logical; should the data be scaled?}

\item{sign_correction}{logical; should \link{signCorrectionPCA} be called?}

\item{return_data}{logical; should data be returned?}

\item{trace}{logical; should progress be printed?}
}
\value{
A list with scores and loadings.
}
\description{
Group-wise sparse principal component analysis
}
