% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupSparseEigen.R
\name{groupSparseEigen}
\alias{groupSparseEigen}
\title{Group-wise sparse spectral decomposition of a matrix}
\usage{
groupSparseEigen(cv, groups, trace = FALSE)
}
\arguments{
\item{cv}{covariance matrix;}

\item{groups}{boolean matrix; affiliation matrix. Does the variable in row i pertain to the group in column j?}

\item{trace}{logical; should progress be printed?}
}
\value{
A list containing the sparse eigenvector and eigenvalues.
}
\description{
Computes group-wise sparse eigenvalues and eigenvectors of a matrix.
}
\details{
For each group defined in the \code{groups} parameter,
variables not pertaining to the group see their covariance set to zero.
Eigendecomposition is then performed on that matrix.
}
\seealso{
\link[base]{eigen}
}
