% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_to_count.R
\name{sleuth_to_count}
\alias{sleuth_to_count}
\title{Convert a sleuth object to count}
\usage{
sleuth_to_count(
  obj,
  log = TRUE,
  normalize = TRUE,
  pc = 0.5,
  aggregation_column = "wb"
)
}
\arguments{
\item{obj}{\code{sleuth} object.}

\item{log}{bool; should natural log be returned?}

\item{normalize}{bool; should normalized counts be returned}

\item{pc}{numeric; pseudo-counts}

\item{aggregation_column}{character; aggregation column}
}
\value{
a matrix which contains a matrix of target_ids and transcript (or gene) expression in read counts.
}
\description{
Convert a sleuth object to a count matrix with the condition names.
}
